Implemented RenderManager header files
The rendermanager skeleton is done, now to flesh it out. IMPORTANT!!!
You must copy visual studio requires that we copy SDL2.lib and
SDL2main.lib to the same folder that the SquadIO .h and .cpp files are
in.
Still need to implement the tinyurl library. will provide instructions
on it when finished.