Uses textures instead of surfaces, added getter/setter methods +more
I annotated a lot of the code with NOTE: and TODO: for further insight
as to what specific lines do and how we can use them as we develop
further. Left some of the old code that used surfaces just commented out
to refer back to in case our texture implementation runs into any bugs.
Mega Props to Ricardo as well!!!