Included Box2D
Adhered to the file structure that was established with header files and
source files being separated out. Quite a bitch of a process to reformat
all the includes in all the files. Haven't implemented any of it yet,
but I ran a very brief test of including the library and compiling with
SceneObject given a b2Body property and it works so far.