Trimmed fat, changed how network players are controlled
As a group we should discuss how we want recieved packets to be
processed. As is, recieved packets are stored in a queue. I think it
would be nice to have them stored in a data structure that allows random
access, since the entire queue is emptied (processed) every update call
in the game loop. It would allow the nonplayer characters to pick out
their info from the 'queue' without having to sort it out in an n^2 loop
in the game loop